Update setup.py to correctly build using cx_Freeze on Windows amd64
Currently cx_Freeze is missing a number of files in numpy, scipy and nmrglue
and not copying the .pyd files into the build folder. To get around this
this setup.py collects *all* .pyd for these mdoules and copies them over.
This results in bigger .msi distributions but they work.

Remove this code when cx_Freeze is able to find the files correctly.

https://bitbucket.org/anthony_tuininga/cx_freeze/issue/43/import-errors-when-using-cx_freeze-with