Adds a style manager module to handle assignment of line/marker styles and colours.
This initial implementation is non-interactive, but has the backend to support defined
colour assignments in future. Presently this watches for classes and assigns a unique style
to each. These are persisted in the current session and will ensure figures all remain
consistent across the different plots. An API is provided for plots to access the attributes
and use as neccessary, the full set being available as ls.kwargs, or ls.line_kwargs, or
ls.marker_kwargs for the line and marker specific kwargs.

The next step in implementation is a UI to allow definition of class-name matching to
assign specific colours and marker styles.