Implement resource-checking for plugins (MATLAB, R)
Plugins now implement a basic check for resource requirements and tools
for plugins without required resources are dimmed in the toolbox. A more
elegant approach to this would be nice in future but it works.

Resource handlers manage the interface between plugins and the depended
resources (mainly MATLAB) ensuring consistent single-thread function,
consistent PATH and auto-restart on killing.