Removing dependencies on foundation etc. which are severely slowing down WebView loads.
There is a consistent slowdown/hang when loading tools on Windows. Tracing this showed nothing,
but it turns out it's due to the webview creation which all depend on base.html, which imports
JQuery and the bootstrap css. Mostly these aren't used (if ever) so removing. Working around
with custom JS has to be better than this.

Later versions of Qt are going to move the browser into a seperate process so it won't block the UI,
and the performance may improve. So, until then.