Implementing apps as Ipython Notebook scripts + Data as pandas
To minimise the development requirement for Pathomx these changes re-implement a number of plugins as IPython notebooks
and convert these to using pandas dataframes for data objects. This is a first commit in a process to convert all of
the plugins to this format.

At present plugins are still defined using Python classes, but it is feasible to replace this with a XML config
file (including definitions for config listings etc.) in future. The ability for users to drag-drop IPython
notebooks into the workspace and auto-create a plugin wrapper is planned: will need the ability to config inputs/outputs.