Modify to add code editor, and integrated config/workflow/tool panels.
This adds a code editor (via QutePart; support for PyQt5 in development with the author)
and changes the kernel to an inprocess kernel with direct variable passing. This greatly
simplifies the handling of data results, etc.

Tools can be edited in place to change how they behave. The file format/load-save will need
to be updated to include the tool code (and about); and a config method developed to handle
update hooks for UI, etc. this will take a while.

Errors are currently ignored but should be highlighted in the editor +/- tooltip/breakpoint.