Implementation of more intelligent Queue execution/scheduling
Initial commit implementing an improved Queue execution and scheduling
approach. By pre-calculating the execution tree and batch executing
the resulting runs on the kernels, this avoids the round-trip signal
through the UI which slows down tool->tool execution.

Each tool is given an Execute object which handles the push/pull of
variables and execution of code. These are assembled into self-contained
Task objects that are linear execution paths, without need for external
variables/etc. These are then passed to the Runner kernel and
handled independently.

Scheduling and forking is incomplete in this commit, and a job listing UI
is to follow to summarise the total progress on the current kernel and
allow start-stop-abort of individually running jobs.